Clarify the role of `main` in this lab
@J-Ziegler documented a significant source of confusion on this lab
last fall, where parts of the write-up suggested you didn't need to care
about the `main.c` files, but you in fact had to make changes there in
order to address all the memory leaks and receive full points.

I've updated the write-up in an effort to hopefully improve that, making
the role of the `main` files clearer.