Rationalize const string pointers
We're now getting issues because of changes to the ISO
C standards about the way that constant strings are
handled. I fixed those by essentially making everthing
a `char const *` when appropriate.

I also changed `char* x` to `char *x` everywhere that
it made sense. This seems to be more common in C land
and certainly is more reasonable if you're declaring
two or more variables on the same line, e.g.,

```
char *x, *y;
```

Lots of people don't recommend declaring multiple
variables on the same line (which I understand), but
since it's legal (and often done) using the syntax
that makes sense for that case seems reasonable.