Draft of a workflow for testing `disemvowel`
This is probably a bit of a mess. It downloads and installs `gtest`, which is likely to be expensive. I've added some caching, but I don't think it's actually right or complete; I think I need to add some sort of conditional to the step that installs `gtest` before the caching will actually matter.

This *only* tests `disemvowel`, so I'll need something like this for the other parts of the lab, as well as valgrind.

Maybe the way to approach this is to have one action that _builds_ everything and then using caching to save those artifacts across tests?