Add GitHub Actions tests for this lab (#5)
This adds a set of six tests in GitHub Actions (with associated badges):

- `palindrome-gtest` runs the GTest tests on the Palindrome problem;
  the badge is red if any of those tests fail.
- `palindrome-main-valgrind` runs `valgrind` on the `main` program
  for the Palindrome problem; the badge is red if there are any
  memory leaks in running `valgrind` on that executable.
- `palindrome-test-valgrind` runs `valgrind` on the executable
  containing the GTest tests; the badge is red if there are any
  memory leaks in running `valgrind` on that executable.

There are then the same three tests, but for the Disemvowel
problem.

My implementation here is somewhat brutish and slow to run. In particular I think we could do something clever with caching to avoid downloading and building `gtest` over and over, and downloading `valgrind` and its dependencies multiple times. I've including caching in the actions, but I don't think what I have _uses_ the caching in any way because I don't test that things are already present. I need to get this out for now, though, so I'll make that an issue and move on.