Add `which` call to figure out where `valgrind` is installed
I'm not sure we can gracefully cache `valgrind` as it appears to depend on a lot of libraries, and I don't want to try to add all that to the cache. (If nothing else it'll be really fragile).

I suspect that what we want to make a separate GitHub Action tool specifically for `valgrind`, but that sounds like work I don't have time for right now.