Avoid installing gtest if it's in the cache (#10)
I'm hoping that this will actually use the cache (that I think was being ignored) to avoid redownloading and installing the gtest library over and over in our GitHub Actions, thus speeding them up.

I don't honestly think that the caching works yet, but it's possible that the changes have to be merged into main for us to know for sure, so I'm going to go ahead and merge this in, hoping for the best. There are some useful/important changes here (e.g., avoiding duplicate builds because of both push and pull-request), so it's worth having these changes regardless.

* Avoid installing gtest if it's in the cache

I'm hoping that this will actually _use_ the cache (that I think was being ignored) to avoid redownloading and installing the `gtest` library over and over in our GitHub Actions, thus speeding them up.

* Drop `pull_request` from GitHub Actions triggers

Having both `push` and `pull_request` essentially doubles the number of GitHub Actions runs, which is really expensive.

* Change all palindrome actions to use `~/lib`

* Remove `gtest` build from action that doesn't need it`

The `disemvowel_main_valgrind` and `palindrome_main_valgrind.yml` actions don't actually use `gtest` anywhere, so this removes building that from the process.