Add include files
It turns out that to compile code that uses the `gtest` library, we also need to have the include files, which weren't being cached before.

Hopefully this adds them to the cache. I've also added some other bits to try to figure out where things are in case this doesn't work on its own.