Fix the `gtest` caching (#11)
I think this may get the `gtest` caching to actually work. I've had to make a bunch of changes, but I'm hoping this pulls everything together to actually get the caching to work, which will hopefully speed up the tests quite a lot.

* Break up huge `gmake` build statement

This breaks that huge thing up into a sequence of separate `run` statements. I think we'll get feedback on these individually as a result?

* Add missing `.outputs` clause

I'd left `.output`out from the `if` statement, which is hopefully why it seemed to always be ignoring the fact that the `gtest` library was cached.

* Add `id` to GitHub Actions cache step

I really need to read the instructions more carefully. The caching step needs to have an `id`, which is then used in the `if` step to find out if we had a cache hit.

* Add include files

It turns out that to compile code that uses the `gtest` library, we also need to have the include files, which weren't being cached before. Hopefully this adds them to the cache.

I also added a `-I` flag to the `g++` builds to include from the cached directory.